Dodavanje entiteta glasanje, promoMaterijal, PromoMaterijalRepo, PromoMaterijalController

+++ IzvorniKod/backend/src/main/java/opp/dao/FotoMaterijalRepo.java
@@ -4,6 +4,6 @@ import opp.domain.FotoMaterijal;
 import opp.domain.Konferencija;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FotoMaterijalRepo extends JpaRepository<FotoMaterijal, Integer> {
+public interface FotoMaterijalRepo extends JpaRepository<FotoMaterijal, Long> {
     //U <> upisujemo tip entiteta i tip ID-a
 }
